+++ src/main/java/org/silly/rats/shop/order/details/OrderStatusRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 public interface OrderStatusRepository
-	extends JpaRepository<OrderStatus, Integer> {
+	extends JpaRepository<OrderStatus, Byte> {
 	OrderStatus findByName(String name);
 
 	@Query("SELECT s.id FROM OrderStatus s WHERE s.name = ?1")
-	Integer getIdByName(String name);
+	Byte getIdByName(String name);
 }
